+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Improved GetDatabase API call caching for Firestore function deployments. (#8681)
 - Increased timeout for linking CloudSQL instances to Data Connect.
 - Fixed issue where `firebase use --add` didn't correctly set the active project. (#8694)
+- Correctly support '--database' flag in `firestore:delete`. (#6753)